Combined all apply method bvas and tests into Game and reordered tests based on enum order
